Fixing go linting issues.
