+++ schema/helpers/base/cliCommand.go
@@ -6,7 +6,6 @@
 package base
 
 import (
-	"errors"
 	"fmt"
 	"github.com/cosmos/cosmos-sdk/client/context"
 	"github.com/cosmos/cosmos-sdk/client/flags"
@@ -39,9 +38,9 @@ func (cliCommand cliCommand) ReadInt64(cliFlag helpers.CLIFlag) int64 {
 			}
 		}
 	default:
-		panic(errors.New(fmt.Sprintf("Flag %v not an int64 flag, Flag type: %T, ", cliFlag.GetName(), cliFlag.GetValue())))
+		panic(fmt.Errorf("flag %v not an int64 flag, Flag type: %T, ", cliFlag.GetName(), cliFlag.GetValue()))
 	}
-	panic(errors.New(fmt.Sprintf("Uregistered flag %v type %T", cliFlag.GetName(), cliFlag.GetValue())))
+	panic(fmt.Errorf("uregistered flag %v type %T", cliFlag.GetName(), cliFlag.GetValue()))
 }
 
 func (cliCommand cliCommand) ReadInt(cliFlag helpers.CLIFlag) int {
@@ -53,9 +52,9 @@ func (cliCommand cliCommand) ReadInt(cliFlag helpers.CLIFlag) int {
 			}
 		}
 	default:
-		panic(errors.New(fmt.Sprintf("Flag %v not an int flag, Flag type: %T, ", cliFlag.GetName(), cliFlag.GetValue())))
+		panic(fmt.Errorf("flag %v not an int flag, Flag type: %T, ", cliFlag.GetName(), cliFlag.GetValue()))
 	}
-	panic(errors.New(fmt.Sprintf("Uregistered flag %v type %T", cliFlag.GetName(), cliFlag.GetValue())))
+	panic(fmt.Errorf("uregistered flag %v type %T", cliFlag.GetName(), cliFlag.GetValue()))
 }
 
 func (cliCommand cliCommand) ReadBool(cliFlag helpers.CLIFlag) bool {
@@ -67,9 +66,9 @@ func (cliCommand cliCommand) ReadBool(cliFlag helpers.CLIFlag) bool {
 			}
 		}
 	default:
-		panic(errors.New(fmt.Sprintf("Falg %v not an bool flag, Flag type: %T, ", cliFlag.GetName(), cliFlag.GetValue())))
+		panic(fmt.Errorf("flag %v not an bool flag, Flag type: %T, ", cliFlag.GetName(), cliFlag.GetValue()))
 	}
-	panic(errors.New(fmt.Sprintf("Uregistered flag %v type %T", cliFlag.GetName(), cliFlag.GetValue())))
+	panic(fmt.Errorf("uregistered flag %v type %T", cliFlag.GetName(), cliFlag.GetValue()))
 }
 
 func (cliCommand cliCommand) ReadString(cliFlag helpers.CLIFlag) string {
@@ -81,9 +80,9 @@ func (cliCommand cliCommand) ReadString(cliFlag helpers.CLIFlag) string {
 			}
 		}
 	default:
-		panic(errors.New(fmt.Sprintf("Falg %v not an string flag, Flag type: %T, ", cliFlag.GetName(), cliFlag.GetValue())))
+		panic(fmt.Errorf("falg %v not an string flag, Flag type: %T, ", cliFlag.GetName(), cliFlag.GetValue()))
 	}
-	panic(errors.New(fmt.Sprintf("Uregistered flag %v type %T", cliFlag.GetName(), cliFlag.GetValue())))
+	panic(fmt.Errorf("uregistered flag %v type %T", cliFlag.GetName(), cliFlag.GetValue()))
 }
 
 func (cliCommand cliCommand) ReadBaseReq(cliContext context.CLIContext) rest.BaseReq {
