+++ utilities/rest/sign/request_test.go
@@ -3,22 +3,20 @@
  SPDX-License-Identifier: Apache-2.0
 */
 
-package signTx
+package sign
 
 import (
 	sdkTypes "github.com/cosmos/cosmos-sdk/types"
+	"github.com/cosmos/cosmos-sdk/types/rest"
 	authTypes "github.com/cosmos/cosmos-sdk/x/auth/types"
-	"github.com/persistenceOne/persistenceSDK/constants/errors"
 	"github.com/stretchr/testify/require"
 	"testing"
 )
 
-func Test_SignTx_Response(t *testing.T) {
-
+func Test_SignTx_Request(t *testing.T) {
+	fromAddress := "cosmos1pkkayn066msg6kn33wnl5srhdt3tnu2vzasz9c"
+	testBaseReq := rest.BaseReq{From: fromAddress, ChainID: "test", Fees: sdkTypes.NewCoins()}
 	testFee := authTypes.NewStdFee(12, sdkTypes.NewCoins())
 	testStdTx := authTypes.NewStdTx([]sdkTypes.Msg{}, testFee, []authTypes.StdSignature{}, "")
-	require.Equal(t, response{Success: true, Error: nil, StdTx: testStdTx}, newResponse(testStdTx, nil))
-	testResponse := newResponse(testStdTx, errors.IncorrectFormat)
-	require.Equal(t, false, testResponse.IsSuccessful())
-	require.Equal(t, errors.IncorrectFormat, testResponse.GetError())
+	require.Equal(t, nil, request{BaseRequest: testBaseReq, Type: "type", StdTx: testStdTx}.Validate())
 }
